requirements.txt
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
azure-core==1.24.1
azure-eventhub==5.10.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
fastapi==0.78.0
h11==0.13.0
idna==3.3
packaging==21.3
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.3
requests==2.28.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
uamqp==1.5.3
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.1
